+++ example-usage/sts3x_i2c_example_usage.c
@@ -3,7 +3,7 @@
  *
  * Generator:     sensirion-driver-generator 0.33.0
  * Product:       sts3x
- * Model-Version: 1.0.0
+ * Model-Version: 2.0.0
  */
 /*
  * Copyright (c) 2023, Sensirion AG
@@ -64,7 +64,8 @@ int main(void) {
         printf("error executing start_periodic_measurement(): %i\n", error);
         return error;
     }
-    float a_temperature = 0.0;
+
+    int32_t a_temperature = 0.0;
     uint16_t repetition = 0;
     for (repetition = 0; repetition < 50; repetition++) {
         error = sts3x_blocking_read_measurement(&a_temperature);
@@ -72,12 +73,13 @@ int main(void) {
             printf("error executing blocking_read_measurement(): %i\n", error);
             continue;
         }
-        printf("a_temperature: %.2f\n", a_temperature);
+        printf("a_temperature [milli degC]: %i\n", a_temperature);
     }
 
     error = sts3x_stop_measurement();
     if (error != NO_ERROR) {
         return error;
     }
+
     return NO_ERROR;
 }