+++ CHANGELOG.md
@@ -7,7 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.0.0] - 2023-11-14
+
+### CHANGED
+
+- Remove usage of float.
+  sht3x_measure_single_shot and sht3x_blocking_read_measurement now return a int32_t,
+  conversion from ticks to physical value uses an approximation and the returned
+  temperature is in milli degrees celsius and humidity in milli %.
+
+
 ## [1.0.0] - 2023-10-27
 
 Initial release
 
+[unreleased] https://github.com/Sensirion/embedded-i2c-sts3x/compare/2.0.0...HEAD
+[2.0.0] https://github.com/Sensirion/embedded-i2c-sts3x/compare/1.0.0...2.0.0
+[1.0.0] https://github.com/Sensirion/embedded-i2c-sts3x/releases/tag/1.0.0